anomaly_detector/requirements.txt
kafka-python
pyod
pandas
scikit-learn
prometheus-client
